UI: Scan QR on home screen
Scan QR Code on home screen instead of a separate QR page.

Rationale:
When visiting shops, scanning QRs is still the main mode of
check-in. Thus, I would like to be able to scan QRs without
any additional UI interaction after opening the app.

Access to history and favorite is still available by scrolling
up the page.

Changes:
 - Remove QR scan page
 - Add QR scan feature to the main page
 - QR button now redirects to the main apge
